szechuan-sauce: Single Source of Truth — one home for the rate-limit probe knob
The env var name, default, floor and resolver body for
PICKLE_RATE_LIMIT_PROBE_INTERVAL_MS existed verbatim in both mux-runner.ts
and microverse-runner.ts, together with the probe protocol constants
(timeout, log filename, prompt). One documented operator knob had two
compiled definitions, so a change to either default or floor would have
left the two waiting runners disagreeing with nothing to notice.

Neither runner could own it (microverse-runner imports from mux-runner, so
the reverse import is a cycle), so the declaration moves to
services/pickle-utils.ts, a module both already import. The parseInt parse
is carried over verbatim rather than tightened to match the sibling
resolveTierStallThresholdMs: that would be a behaviour change to both
runners wearing the clothes of a move.
