+++ extension/bin/mux-runner.js
@@ -3,7 +3,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
 import { spawn, spawnSync, execFileSync } from 'child_process';
-import { printMinimalPanel, Style, formatTime, getExtensionRoot, getDataRoot, formatLocalDateKey, buildHandoffSummary, sleep, writeStateFile, markTicketDone, markTicketSkipped, markTicketWithStatus as writeTicketStatus, collectTickets, getTicketStatus, runCmd, safeErrorMessage, ensureMonitorWindow, displayMacNotification, parseTicketFrontmatter, getTicketTierBudgetWithOverrides, readFrontmatterField, upsertFrontmatterField, ticketFilePath, VALID_TICKET_COMPLEXITY_TIERS, TIER_LIFECYCLE, composeManagerPromptFromSkill, resolveWorkerTestGateTimeoutMs, scrubGateEnv, resolveCommandTemplate, resolveManagerPromptPath, loadPickleSettingsBag, resolveHardeningSettings, resolveCodegraphSettings, resolveRateLimitSettings, DEFAULT_MAX_PARK_MINUTES } from '../services/pickle-utils.js';
+import { printMinimalPanel, Style, formatTime, getExtensionRoot, getDataRoot, formatLocalDateKey, buildHandoffSummary, sleep, writeStateFile, markTicketDone, markTicketSkipped, markTicketWithStatus as writeTicketStatus, collectTickets, getTicketStatus, runCmd, safeErrorMessage, ensureMonitorWindow, displayMacNotification, parseTicketFrontmatter, getTicketTierBudgetWithOverrides, readFrontmatterField, upsertFrontmatterField, ticketFilePath, VALID_TICKET_COMPLEXITY_TIERS, TIER_LIFECYCLE, composeManagerPromptFromSkill, resolveWorkerTestGateTimeoutMs, scrubGateEnv, resolveCommandTemplate, resolveManagerPromptPath, loadPickleSettingsBag, resolveHardeningSettings, resolveCodegraphSettings, resolveRateLimitSettings, resolveRateLimitProbeIntervalMs, RATE_LIMIT_PROBE_TIMEOUT_MS, RATE_LIMIT_PROBE_LOG_FILENAME, RATE_LIMIT_PROBE_PROMPT, DEFAULT_MAX_PARK_MINUTES } from '../services/pickle-utils.js';
 import { findMissingPrefixes, requiredTierArtifactPrefixes } from '../services/artifact-validation.js';
 import { PromiseTokens, hasToken, VALID_STEPS, Defaults, classifyExitReason, FALSE_EPIC_THRESHOLD, hasLifecycleArtifact, matchesArtifactPrefix, newestArtifactFile, NO_PROGRESS_FAILURE_REASONS, WORKER_GATE_VERDICT_FIELD, UNBOUNDED_READ_MAX_BUFFER, enumerationCompleted, reportedTestResults } from '../types/index.js';
 import { StateManager, safeDeactivate, finalizeTerminalState, finalizeIfTrulyComplete, recordExitReason, clearExitReason, writeActivityEntry, writeTimeoutStub, schemaVersionDeployDriftMessage, isProcessAlive } from '../services/state-manager.js';
@@ -7717,21 +7717,6 @@ async function waitThroughRateLimit(ctx, resetAtSec, minWaitMs) {
     }
     return { exit: false };
 }
-const RATE_LIMIT_PROBE_TIMEOUT_MS = 120_000;
-const RATE_LIMIT_PROBE_LOG_FILENAME = 'rate_limit_probe.log';
-const RATE_LIMIT_PROBE_PROMPT = 'Reply with exactly: ok';
-/** Same env var microverse-runner.ts's resolveRateLimitProbeIntervalMs reads — one operator knob for
- * "how often does a rate-limit wait re-probe", not two. Duplicated here (not imported) because
- * microverse-runner.ts imports FROM mux-runner.ts, so the reverse import would be circular. */
-const RATE_LIMIT_PROBE_INTERVAL_ENV_VAR = 'PICKLE_RATE_LIMIT_PROBE_INTERVAL_MS';
-const DEFAULT_RATE_LIMIT_PROBE_INTERVAL_MS = 10 * 60 * 1000;
-const MIN_RATE_LIMIT_PROBE_INTERVAL_MS = 60_000;
-function resolveMuxRateLimitProbeIntervalMs(env = process.env) {
-    const raw = parseInt(env[RATE_LIMIT_PROBE_INTERVAL_ENV_VAR] ?? '', 10);
-    if (!Number.isFinite(raw) || raw <= 0)
-        return DEFAULT_RATE_LIMIT_PROBE_INTERVAL_MS;
-    return Math.max(raw, MIN_RATE_LIMIT_PROBE_INTERVAL_MS);
-}
 /** Minimal spawn-and-capture with a timeout — the mux-runner-local twin of
  * microverse-runner.ts's private (unexported) spawnWithClosedStdin. */
 function spawnRateLimitProbe(cmd, args, opts) {
@@ -7821,7 +7806,7 @@ async function probeMuxRateLimitCleared(input) {
 async function waitOutRateLimitParkWithProbe(input, now, sleepFn, parkStartMs, resumeTargetMs) {
     const { statePath, sessionDir, state } = input;
     const probeFn = input.probeApiCleared ?? probeMuxRateLimitCleared;
-    const probeIntervalMs = input.probeIntervalMs ?? resolveMuxRateLimitProbeIntervalMs();
+    const probeIntervalMs = input.probeIntervalMs ?? resolveRateLimitProbeIntervalMs();
     let lastProbeMs = parkStartMs;
     while (now() < resumeTargetMs) {
         await sleepFn(Defaults.RATE_LIMIT_POLL_MS);
